+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Fix issue with `useViewModelCache` that could cause a render after unmount
 * (breaking) Removed `RequestError` from `Endpoint`. Instead `Endpoint` will now throw the error raised by `fetch` directly - either `TypeError` or `AbortError`.
+* Fix bugs with `RelatedViewModelField` and `ManyRelatedViewModelField` where things could break if value was `null` or `[]`
 
 ## [0.0.11] - 2021-03-01